task: moved naverror outside
